+++ src/index.js
@@ -2,7 +2,7 @@ import React, { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./styles.css";
 
-import App from "./App";
+import App from "./Toodle";
 
 const root = createRoot(document.getElementById("root"));
 root.render(
