+++ src/styles.css
@@ -57,19 +57,29 @@ body {
 }
 
 .square {
-  background: #fff;
-  border: 1px solid #999;
-  float: left;
+  border-radius: 5px;
+  margin: 2px;
+  width: 40px;
+  height: 40px;
   font-size: 24px;
   font-weight: bold;
-  line-height: 34px;
-  height: 34px;
-  margin-right: -1px;
-  margin-top: -1px;
-  padding: 0;
-  text-align: center;
-  width: 34px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: 1px solid #999;
 }
+
+.board {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.board-row {
+  display: flex;
+  justify-content: center;
+}
+
 
 .board-row:after {
   clear: both;
